+++ test/io.jl
@@ -6,7 +6,7 @@
         @test !PMD.ismultinetwork(base_network)
         @test PMD.iseng(network) && PMD.iseng(base_network)
 
-        @test length(network["nw"]) == 3
+        @test length(network["nw"]) == 5
 
         @test network["nw"]["1"]["switch"]["671700"]["state"] == PMD.CLOSED
         @test network["nw"]["1"]["switch"]["671700"]["dispatchable"] == PMD.YES
@@ -15,7 +15,7 @@
 
     @testset "test events parsing" begin
         raw_events = parse_events("../test/data/events.json")
-        @test length(raw_events) == 3
+        @test length(raw_events) == 7
 
         events = parse_events(raw_events, network)
         @test isa(events, Dict) && length(events) == 2
@@ -27,13 +27,10 @@
         @test _network["nw"]["1"]["switch"]["671700"]["state"] == PMD.OPEN
         @test _network["nw"]["1"]["switch"]["671700"]["dispatchable"] == PMD.NO
         @test _network["nw"]["1"]["switch"]["671700"]["status"] == PMD.ENABLED
-        @test _network["nw"]["1"]["switch"]["671692"]["dispatchable"] == PMD.NO
 
         @test _network["nw"]["2"]["switch"]["671700"]["dispatchable"] == PMD.YES
-        @test _network["nw"]["2"]["switch"]["671692"]["dispatchable"] == PMD.NO
 
         @test _network["nw"]["3"]["switch"]["671700"]["dispatchable"] == PMD.YES
-        @test _network["nw"]["3"]["switch"]["671692"]["dispatchable"] == PMD.NO
     end
 
     @testset "test settings parsing" begin